argoslabs/excel/image/requirements.txt
# pip dependent packages

# default for alabs
alabs.common

# for git html
pywin32
Pillow